feat(EwSigner): add fromEthersSigner
Also removed fromEthersProvider as it is not necessary to have.
It makes more sense to create from signer as we are creating a EW"Signer"

Also, added check that signer is actually available if creating from EIP1193
This is kinda of the issue that the EKC team had where they had a provider
that wasn't actually connected to a signer. So better to check early.